backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
elasticsearch
sentence-transformers
anthropic
